further refactor work
