+++ augmentation/Augmenting_Features.py
@@ -1,194 +1,10 @@
+# Augmenting_Features.py
 import cv2
 import numpy as np
 import os
-from landmarking import load_feature_landmarks, resize_and_overlay_feature
+from augmentation import resize_nose, create_eye_mask, create_nose_mask, make_img
+from landmarking import load_feature_landmarks
 from utils import get_dir
-
-'''
-Name: Rhiannon Barber
-Date: January 25, 2024
-
-This file contains the code for augmenting the facial features (nose, and eyes currently as of 1/25).
-
-The augmentation is done to a hardcoded scale value, that needs to be changed to work with the std
-deviation found in the Facial_Landmarking_Math.py file. 
-
-As of 1/25 I have successfully blurred in the nose region for both larger and smaller scale values. 
-
-Currently working on the eyes, though the code is not included in this file since it is very messy and not 
-working correctly.  
-
-In line comments have more detail on how certain things are working. 
-'''
-
-'''
-This function loads the land markings found from the .npy file
-returns the x,y coordinates from the data with are col 4 & 5.
-'''
-
-
-def load_feature_landmarks(facial_features, image_id, feature_to_int, feature_name):
-    feature_landmarks = facial_features[
-        (facial_features[:, 0] == image_id) & (facial_features[:, 1] == feature_to_int[feature_name])]
-    return feature_landmarks[:, 3:5]
-
-
-'''
- This function is responsible for resizing the nose.
- Gets the coordinates from the .npy file, and finds the feature named 'nose'
- returns the resized nose, with a margin factor to capture the nostrils
- region is overlaid on the image.
-'''
-
-
-def resize_nose(img, facial_features, image_id, feature_to_int, scale_factor):
-    nose_landmarks = load_feature_landmarks(facial_features, image_id, feature_to_int, 'nose')
-    return resize_and_overlay_feature(img, nose_landmarks, scale_factor, width_margin_factor=0.6,
-                                      height_margin_factor=0.7)
-
-
-'''
-Similar to resize nose, but for eyes.
-returns the resized eyes, with margin to capture eye lashes.
-region is overlaid on the image.
-'''
-
-
-def resize_eyes(img, facial_features, image_id, feature_to_int, scale_factor):
-    eye_landmarks = load_feature_landmarks(facial_features, image_id, feature_to_int, 'eyes')
-    return resize_and_overlay_feature(img, eye_landmarks, scale_factor, width_margin_factor=0.25,
-                                      height_margin_factor=0.4)
-
-
-'''
-This function is what is returned in the resize method. 
-It's responsible for resizing the feature points, based on the scale factor, and the margin points. 
-it is then overlaid onto the original image, and the 
-'''
-
-
-def resize_and_overlay_feature(img, feature_points, scale_factor, width_margin_factor, height_margin_factor):
-    # Calculates the bounding box for the region, based on margin and landmarks.
-    x_min, y_min = np.min(feature_points, axis=0).astype(int)
-    x_max, y_max = np.max(feature_points, axis=0).astype(int)
-    width_margin = int((x_max - x_min) * width_margin_factor)
-    height_margin = int((y_max - y_min) * height_margin_factor)
-    x_min = max(x_min - width_margin, 0)
-    y_min = max(y_min - height_margin, 0)
-    x_max = min(x_max + width_margin, img.shape[1])
-    y_max = min(y_max + height_margin, img.shape[0])
-    feature_region = img[y_min:y_max, x_min:x_max]
-    feature_width = x_max - x_min
-    feature_height = y_max - y_min
-    scaled_width = int(feature_width * scale_factor)
-    scaled_height = int(feature_height * scale_factor)
-
-    # cv2.resize takes the image, the scale to resize, and the interpolation: I chose linear since it was the
-    # simplest approach to scaling the region with good results.
-    scaled_feature = cv2.resize(feature_region, (scaled_width, scaled_height), interpolation=cv2.INTER_LINEAR)
-
-    # Calculates the overlay starting horizontal margin for where the resized feature should be placed.
-    # Discarding the fractional part to ensure whole integers for coordinates.
-    overlay_start_x = x_min + (feature_width - scaled_width) // 2
-
-    # Calculates the overlay starting vertical margin for where the resized feature should be placed.
-    # Discarding the fractional part to ensure whole integers for coordinates.
-    overlay_start_y = y_min + (feature_height - scaled_height) // 2
-    img[overlay_start_y:overlay_start_y + scaled_height,
-    overlay_start_x:overlay_start_x + scaled_width] = scaled_feature
-    return img
-
-
-'''
-This function creates a mask where the area around the nose is white (255), to indicate the ROI. 
-The rest of the mask is black, it can then be used in the blending.
-The mask uses the land mark points, and margin to define the region of the mask. 
-'''
-
-
-def create_nose_mask(img, feature_points, width_margin_factor, height_margin_factor):
-    mask = np.zeros_like(img)
-    x_min, y_min = np.min(feature_points, axis=0).astype(int)
-    x_max, y_max = np.max(feature_points, axis=0).astype(int)
-    width_margin = int((x_max - x_min) * width_margin_factor)
-    height_margin = int((y_max - y_min) * height_margin_factor)
-    x_min = max(x_min - width_margin, 0)
-    y_min = max(y_min - height_margin, 0)
-    x_max = min(x_max + width_margin, img.shape[1])
-    y_max = min(y_max + height_margin, img.shape[0])
-    mask[y_min:y_max, x_min:x_max] = 255
-    return mask
-
-
-'''
-This function creates a mask where the area around the mask is white, to indicate the ROI. 
-The rest of the mask is black, it can then be used in the blending.
-The mask uses the land mark points to define the region of the mask. 
-'''
-
-
-def create_eye_mask(eye_landmarks1, eye_landmarks2, img):
-    # mask 1, makes mask for reduced eye
-    mask = np.zeros((img.shape[0], img.shape[1], 3), dtype=np.float32)
-    cv2.fillConvexPoly(mask, np.int32(eye_landmarks1), (1.0, 1.0, 1.0))
-    cv2.fillConvexPoly(mask, np.int32(eye_landmarks2), (1.0, 1.0, 1.0))
-    mask = 255*np.uint8(mask)
-
-    # Apply close operation to improve mask
-    kernel = cv2.getStructuringElement(cv2.MORPH_RECT, (40, 40))
-    mask = cv2.morphologyEx(mask, cv2.MORPH_CLOSE, kernel)
-    return mask
-
-
-'''
-This function puts the image in the mask and puts an image in the inverse mask. For mask it puts a reduced image 
-of the eyes in the mask. While for mask2 it puts a blended image in the mask2 and 
-the normal image in each ones inverse mask.
-'''
-
-
-def multiply_eye_mask(mask, inverse_mask, eye, face):
-    # Multiply eyes and face by the mask
-    just_eye = cv2.multiply(mask, eye)
-    just_face = cv2.multiply(inverse_mask, face)
-    result = just_face + just_eye  # Add face and eye
-    return result
-
-
-# Calculates inverse mask and reduces size of image. Then multiplies mask and image together.
-def make_img(img, mask, face_landmarks, eyebrow_landmarks, face_img, type_mask):
-    inverse_mask = cv2.bitwise_not(mask)  # Calculate inverse mask
-    # Convert masks to float to perform blending
-    mask = mask.astype(float)/255
-    inverse_mask = inverse_mask.astype(float)/255
-    img = img.astype(np.uint8)  # Apply color mapping for the eyes
-    # Find amount to reduce image
-    face_width = face_landmarks[0, 1]
-    eyebrow_height = eyebrow_landmarks[3, 1]
-    h, w, c = img.shape
-    height_reduced = int((h-eyebrow_height)/32)
-    width_reduced = int((w-face_width)/8)
-    # Convert eyes and face to 0-1 range
-    face = img.copy()
-    head = img.astype(float)/255
-    if type_mask == 'mask':
-        # Resize
-        eye = cv2.copyMakeBorder(img, height_reduced, height_reduced*3, width_reduced, width_reduced, cv2.BORDER_CONSTANT)
-        eye = cv2.resize(eye, (w, h))
-        eye = eye.astype(float)/255
-        # Multiply eyes and face by the mask
-        result = multiply_eye_mask(mask, inverse_mask, eye, face_img)
-    else:
-        # Resize
-        face = cv2.resize(face, (w*2, h*2))
-        face = cv2.copyMakeBorder(face, 0, 0, int(width_reduced*0.3), int(width_reduced*0.3), cv2.BORDER_CONSTANT)
-        face = cv2.resize(face, (w, h))
-        face = cv2.GaussianBlur(face, (0, 0), 8)
-        face = face.astype(float)/255
-        # Multiply eyes and face by the mask
-        result = multiply_eye_mask(mask, inverse_mask, face, head)
-    return result
-
 
 def main():
     facial_features = np.load(get_dir('data/facial_features.npy'))
