+++ augmentation/__init__.py
@@ -1,3 +1,6 @@
-
-
+# __init.py__ for augmentation package
+from .nose import *
+from .eyes import *
+from .make_image import *
+from .resize_overlay import *
 from .Augmenting_Features import *
