+++ database/__init__.py
@@ -1,5 +1,4 @@
 # __init.py__ for database package
-
 from .createuser_page import *
 from .database_page import *
 from .login_page import *
