+++ landmarking/__init__.py
@@ -1,4 +1,5 @@
 # __init.py__ for landmarking package
 
+from .load import *
 from .Facial_Landmarking import *
 from .Facial_Landmarking_Math import *
